Move remote_request from ogds.base to base and rename to dispatch_request.
