+++ opengever/base/transport.py
@@ -1,9 +1,9 @@
 from five import grok
 from opengever.base.exceptions import TransportationError
 from opengever.base.interfaces import IDataCollector
+from opengever.base.request import dispatch_json_request
 from opengever.ogds.base.utils import decode_for_json
 from opengever.ogds.base.utils import encode_after_json
-from opengever.ogds.base.utils import remote_json_request
 from plone.dexterity.interfaces import IDexterityContent
 from plone.dexterity.utils import addContentToContainer
 from plone.dexterity.utils import createContent
@@ -51,7 +51,7 @@ class Transporter(object):
             REQUEST_KEY: jsondata,
             }
 
-        return remote_json_request(
+        return dispatch_json_request(
             target_cid, '@@transporter-receive-object',
             path=container_path, data=request_data)
 
@@ -61,7 +61,7 @@ class Transporter(object):
         *path* is the relative path of the object to its plone site root.
         """
 
-        data = remote_json_request(source_cid,
+        data = dispatch_json_request(source_cid,
                                    '@@transporter-extract-object-json',
                                    path=path)
