+++ opengever/meeting/proposal.py
@@ -1,14 +1,14 @@
 from opengever.base.oguid import Oguid
 from opengever.base.source import DossierPathSourceBinder
-from opengever.base.transport import REQUEST_KEY
-from opengever.base.transport import Transporter
 from opengever.meeting import _
+from opengever.meeting.command import CopyProposalDocumentCommand
+from opengever.meeting.command import CreateSubmittedProposalCommand
 from opengever.meeting.container import ModelContainer
 from opengever.meeting.model.proposal import Proposal as ProposalModel
 from opengever.meeting.workflow import State
 from opengever.meeting.workflow import Transition
 from opengever.meeting.workflow import Workflow
-from opengever.ogds.base.utils import remote_json_request
+from opengever.ogds.base.utils import ogds_service
 from plone import api
 from plone.directives import form
 from z3c.relationfield.schema import RelationChoice
@@ -16,7 +16,6 @@ from z3c.relationfield.schema import RelationList
 from zope import schema
 from zope.interface import implements
 from zope.interface import Interface
-import json
 
 
 class IProposalModel(Interface):
@@ -82,50 +81,6 @@ class Submit(Transition):
             copy_command.execute()
 
 
-class CreateSubmittedProposalCommand(object):
-
-    def __init__(self, proposal):
-        self.proposal = proposal
-        self.submitted_proposal_path = None
-        # XXX NO! NONONONO!
-        self.admin_unit_id = self.proposal.load_model().committee.admin_unit_id
-
-    def execute(self):
-        model = self.proposal.load_model()
-        jsondata = {'committee_oguid': model.committee.oguid.id,
-                    'proposal_oguid': model.oguid.id}
-        request_data = {
-            REQUEST_KEY: json.dumps(jsondata),
-            }
-        response = remote_json_request(
-            self.admin_unit_id, '@@create_submitted_proposal', data=request_data)
-        self.submitted_proposal_path = response['path']
-
-
-class CopyProposalDocumentCommand(object):
-
-    def __init__(self, document, parent_action):
-        self.document = document
-        self.parent_action = parent_action
-
-    def execute(self):
-        assert self.parent_action.submitted_proposal_path
-        target = self.parent_action.submitted_proposal_path
-        OgCopyCommand(self.document, self.parent_action.admin_unit_id, target).run()
-
-
-class OgCopyCommand(object):
-
-    def __init__(self, source, target_admin_unit_id, target_path):
-        self.source = source
-        self.target_path = target_path
-        self.target_admin_unit_id = target_admin_unit_id
-
-    def run(self):
-        Transporter().transport_to(
-            self.source, self.target_admin_unit_id, self.target_path)
-
-
 class ProposalBase(ModelContainer):
 
     workflow = None
@@ -179,6 +134,11 @@ class ProposalBase(ModelContainer):
             return ''
 
         return model.get_searchable_text()
+
+    def get_committee_admin_unit(self):
+        admin_unit_id = self.load_model().committee.admin_unit_id
+        return ogds_service().fetch_admin_unit(admin_unit_id)
+
 
 
 class SubmittedProposal(ProposalBase):
